08/06/2022 Implementando Room

+++ app/src/main/java/com/example/ecomovingapp/localdatabase/Location.kt
@@ -8,10 +8,11 @@ import com.google.gson.Gson
 @Entity
 data class Location (
     @PrimaryKey (autoGenerate = true) val id:Int,
-    @ColumnInfo (name = "latitude") val latitude:Float,
-    @ColumnInfo (name = "longitude") val longitude:Float,
+    @ColumnInfo (name = "latitude") val latitude:Double,
+    @ColumnInfo (name = "longitude") val longitude:Double,
     @ColumnInfo (name = "description") val description:String?
 ){
+
     override fun toString(): String {
         val gson = Gson()
         return gson.toJson(this)
